requirements/development.txt
Django==1.11.5
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-extensions==1.9.1
djangorestframework==3.6.4
Faker==0.7.3
mixer==5.6.6
psycopg2==2.7.3.1
py==1.4.34
pytest==3.2.2
pytest-django==3.1.2
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
sqlparse==0.2.4
